pretraining_loop: better control over buffers
Same changeset as the one applied to training_loop.py

The general idea is to be more careful with the values we keep live
references to, and where we have them reside.
Previously we held references to a bunch of state values that in turn
referenced device (GPU/TPU) buffers, preventing them from being
GCed/free'd/reused, and forcing JAX to allocate more memory than
necessary, causing OOMs, usually at the beginning of the 2nd epoch.

Now we try and keep references only to host memory.